Working prototype1

+++ source/lib/rocprof-sys/library.cpp
@@ -136,6 +136,18 @@ finalization_handler()
     if(get_state() == State::Active) rocprofsys_finalize();
 }
 
+//Tim: Handles attach/detach. This replaces finalization handler if dl in initialized in attach mode. 
+void
+attach_detach_handler()
+{
+    if (get_state() < State::Active)
+    {
+        ROCPROFSYS_VERBOSE_F(1, "ATTACH ACTIVE\n");
+        rocprofsys_init_tooling_hidden();
+        return;
+    }
+    rocprofsys_finalize();
+}
 auto
 ensure_finalization(bool _static_init = false)
 {
@@ -402,13 +414,15 @@ rocprofsys_init_library_hidden()
     } };
 
     ROCPROFSYS_CONDITIONAL_BASIC_PRINT_F(_debug_init, "\n");
+    //Tim: This sets default signal handler to attach_detach_handler.
+    //Setting it here will prevent finalization_handler from being set later.
+    if (tim::get_env("ROCPROFSYS_ATTACH", true))
+    {
+        ROCPROFSYS_VERBOSE_F(1, "Initializing rocprof-sys in attach mode.\n");
+        config::set_signal_handler(&attach_detach_handler);
+    }
 }
 
-// Initialize RCCL if:
-// - postinit=true - so the code doesn't hang at the initialization stage
-// - get_state() >= State::Init - so the code doesn't throw an exception
-// - rccl_initialized=false - so we don't try to initialize RCCL twice
-// - get_use_rcclp()=true - only if the environment is configured to use RCCL
 static void
 rccl_setup(bool postinit)
 {
@@ -455,20 +469,22 @@ rocprofsys_init_tooling_hidden(bool postinit)
 
     ROCPROFSYS_CONDITIONAL_BASIC_PRINT_F(_debug_init, "State is %s...\n",
                                          std::to_string(get_state()).c_str());
-
+    if (_is_attach)
     if(get_state() != State::PreInit || get_state() == State::Init || _once)
     {
         rccl_setup(postinit);
-        return false;
+        if (!tim::get_env("ROCPROFSYS_ATTACH", false) || get_state() >= State::Active || _once) 
+            return false;
     }
     _once = true;
 
     ROCPROFSYS_SCOPED_THREAD_STATE(ThreadState::Internal);
 
-    ROCPROFSYS_CONDITIONAL_THROW(
-        get_state() == State::Init,
-        "%s called after rocprofsys_init_library() was explicitly called",
-        ROCPROFSYS_FUNCTION);
+    //Tim: Allows this function to be called after rocprofsys_init_library_hidden();
+    // ROCPROFSYS_CONDITIONAL_THROW(
+    //     get_state() == State::Init,
+    //     "%s called after rocprofsys_init_library() was explicitly called",
+    //     ROCPROFSYS_FUNCTION);
 
     ROCPROFSYS_CONDITIONAL_BASIC_PRINT_F(get_verbose_env() >= 0,
                                          "Instrumentation mode: %s\n",
@@ -708,6 +724,7 @@ rocprofsys_finalize_hidden(void)
     // disable initialization callback
     threading::remove_callback(&ensure_initialization);
 
+    bool _is_attach = tim::get_env("ROCPROFSYS_ATTACH", false);
     bool _is_child = is_child_process();
 
     set_thread_state(ThreadState::Completed);
@@ -823,7 +840,9 @@ rocprofsys_finalize_hidden(void)
     if(get_use_rocm())
     {
         ROCPROFSYS_VERBOSE_F(1, "Shutting down ROCm...\n");
-        rocprofiler_sdk::shutdown();
+        //Tim: only shutdown rocprofiler_sdk in non-attach mode, otherwise we will run into segfault
+        if (!_is_attach)
+            rocprofiler_sdk::shutdown();
     }
 #endif
 
@@ -1016,7 +1035,10 @@ rocprofsys_finalize_hidden(void)
         { tim::signals::sys_signal::SegFault, tim::signals::sys_signal::Stop },
         [](int) {});
 
-    common::destroy_static_objects();
+    //Tim: Prevent segfault in attach mode. 
+    //TODO: Find out why this sometimes causes a segfault
+    if (!tim::get_env("ROCPROFSYS_ATTACH", false))
+        common::destroy_static_objects();
 }
 
 //======================================================================================//
